almost posted issue with strcat() [1:21pm]
before I look into a c function that would remove a specified string for a source string,
my last commit worked, but was set up with such a way with strcat that:

strcat() made an irreversible change which was 
appending file 1 onto the string.

users/medium/Desktop/${filepath1)

the filepath1 that I woudn't want to have appended by the time the string is needed to perform the same functionality for filepath2, was concatenated to be so.

char desktop_path[100];
I created an array initializer so that the string:
/Users/medium/Desktop

was able to take a pit-stop:

strcat(desktop_path, userpath);

now:

👍 desktop_path: "Users/medium/Desktop"
👍 userpath: "Users/medium/Desktop"

The distinction between these two equivalent (for now) strings:
desktop_path would remain as is with no further concatenation. This is done because the opening and merging of file 2 relies on that string being unaltered.
userpath would go on to have file1 concatenated onto it.

fun little puzzle to solve.